refactor: LegacyHandlerMapping 클래스 제거
